rename types folder
